[refactor] #12 global 폴더 변경
